Move AzleFnDecl into canister method module
