+++ src/compiler/typescript_to_rust/azle_generate/src/canister_method/errors/duplicate_method_types.rs
@@ -5,7 +5,7 @@ use annotate_snippets::{
 use cdk_framework::act::node::canister_method::CanisterMethodType;
 use swc_common::{source_map::Pos, Span};
 
-use crate::ts_ast::{source_map::GetSourceFileInfo, AzleFnDecl};
+use crate::{canister_method::AzleFnDecl, ts_ast::source_map::GetSourceFileInfo};
 
 pub fn build_duplicate_method_types_error_message_from_azle_fn_decl(
     azle_fn_decls: Vec<AzleFnDecl>,
