+++ src/compiler/typescript_to_rust/azle_generate/src/canister_method/query_and_update/shared/body.rs
@@ -4,7 +4,7 @@ use swc_ecma_ast::{
     TsType,
 };
 
-use crate::{canister_method::rust, ts_ast::AzleFnDecl};
+use crate::canister_method::{rust, AzleFnDecl};
 
 pub use super::build_canister_methods::BuildCanisterMethods;
