+++ src/compiler/typescript_to_rust/azle_generate/src/ts_ast/mod.rs
@@ -5,14 +5,12 @@ use swc_ecma_parser::{lexer::Lexer, Parser, StringInput, Syntax, TsConfig};
 
 use crate::ts_ast::source_map::SourceMapped;
 
-pub use azle_fn_decl::AzleFnDecl;
 pub use azle_functions_and_methods::FunctionAndMethodTypeHelperMethods;
 pub use azle_type::{AzleArrayType, AzleFnOrConstructorType};
 pub use azle_type_alias_decl::AzleTypeAliasDecl;
 pub use traits::GetName;
 pub use traits::{GetSourceText, GetTsType};
 
-mod azle_fn_decl;
 mod azle_functions_and_methods;
 mod ident;
 mod ts_fn_param;
